ToadsTreasureHunt update switch
Sorry for all the updates, ignore all the others if you like and make this one the initial release if that's easier. Last configuration relied too much on luck with the event squares I chose, so I switched them out and added a better use of the switch. Thanks for all you guys do putting this site together.