Toad's Treasure Hunt 4 switch states
Added two more switch states, so now The map alternates between the gold events switching spots with the four for sale plots on the outer edges and then the ones on the inside paths. Deleted the change log and labeled this yaml as initial release since technically the map has still yet to be released, so makes sense to still say V1.