Add files via upload
The description was kind of overly wordy and the quotation marks were causing the end to be cut off. So I replaced it with this shorter sentence: A haunted, shapeshifting labyrinth of King Boo's own design, meant to hide his treasure from the living. The passageway in contains a message...Trespassers beware! It waits for you behind door 3, then 2, then 1, now the haunt has just begun