New Map: POW Block
Medium sized map. Switch has two functions. 1. swaps the properties that are on the outer ring with the inner ring, and 2. every other press the bank switches, and stock broker rotate counter clockwise. Together this combo makes it so every district a chance to pin a suit and have a square located near the bank throughout the duration of a match.