KBHH hotfix
reverts the total board states back to 48. Doing this made the background glitch go away on my pc. The 6 frb states that contained too much stutter before have been replaced with a doubling of 13abc and 10abc in the yaml reducing the total frbs to 42, but keeping the number of switch states at 48.