Fixes a possible soft lock
Fixes a soft lock that could occur if the player travels from the bank past the club to land on the switch, which would not allow the player to move past the club again. I doubled the club suit so that the player is able to pass through that direction in the event this roll happens.